refactor: rename main to app
