+++ app/src/main/java/com/cyanlch/howgoods/HgApplication.kt
@@ -4,4 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class SApplication : Application()
+class HgApplication : Application()
